refactor: 정적 리소스 정보 관리 클래스 분리
